Examen p2

+++ src/app/components/inventario/inventario.component.ts
@@ -14,13 +14,13 @@ import { Router } from '@angular/router';
   styleUrl: './inventario.component.css'
 })
 export class InventarioComponent {
-  productos$: Observable<Producto[]>; // Observables para productos
 
   nuevoProducto: Producto = { id: 0, nombre: '', precio: 0, imagen: '' };
 
-  constructor(public inventarioService: InventarioService, private router: Router) {
-    this.productos$ = this.inventarioService.productos$; // Asignación dentro del constructor
-  }
+   constructor(
+    public inventarioService: InventarioService,
+    private router: Router
+  ) {}
 
   agregarProducto() {
     if (!this.nuevoProducto.nombre || this.nuevoProducto.precio <= 0) {
@@ -28,29 +28,14 @@ export class InventarioComponent {
       return;
     }
 
-    // Verificar si la ID ya existe en la lista de productos
-    this.productos$.subscribe(productos => {
-      if (productos.some(producto => producto.id === this.nuevoProducto.id)) {
-        alert("La ID del producto ya existe. Por favor, elija otra.");
-        return;
-      } else {
-        const nuevoProducto: Producto = {
-          id: this.nuevoProducto.id,
-          nombre: this.nuevoProducto.nombre,
-          precio: this.nuevoProducto.precio,
-          imagen: this.nuevoProducto.imagen || `asset/noimagen.jpg`, // Imagen por defecto  
-        };
-
-        this.inventarioService.agregarProducto(nuevoProducto);
-
-        // Reiniciar formulario
-        this.nuevoProducto = { id: 0, nombre: '', precio: 0, imagen: '' };
-      }
-    });
+    this.inventarioService.agregarProducto(this.nuevoProducto);
+    this.nuevoProducto = { id: 0, nombre: '', precio: 0, imagen: '' };
   }
 
   eliminarProducto(id: number) {
-    this.inventarioService.eliminarProducto(id);
+    if (confirm('¿Está seguro de que desea eliminar este producto?')) {
+      this.inventarioService.eliminarProducto(id);
+    }
   }
 
   descargarXML() {
